backend/requirements.txt
# Web Framework
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Banco de Dados
supabase>=1.0.3

# Segurança
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1

# Integração com APIs
httpx>=0.25.0
aiohttp>=3.8.6
tenacity>=8.2.3

# IA e Análise de Dados
anthropic>=0.49.0
pandas>=2.1.1
numpy>=1.26.0
loguru>=0.7.2

# Utilitários
beautifulsoup4>=4.12.2
colorama>=0.4.6

# Testes
pytest==7.4.2
pytest-asyncio==0.21.1
